# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "FertiScan", "version": "0.1.0" },
  "paths": {
...
    "/api/v1/labels": {
      "post": {
        "tags": ["labels"],
        "summary": "Create Label",
        "description": "Create a new label.",
        "operationId": "labels_create_label",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LabelCreate" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LabelCreated" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["labels"],
        "summary": "Read Labels",
        "description": "List labels with optional filters and sorting.",
        "operationId": "labels_read_labels",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "review_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                { "$ref": "#/components/schemas/ReviewStatus" },
                { "type": "null" }
              ],
              "description": "Filter by review status",
              "title": "Review Status"
            },
            "description": "Filter by review status"
          },
          {
            "name": "unlinked",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "boolean" }, { "type": "null" }],
              "description": "Filter labels not linked to a product (product_id is null)",
              "title": "Unlinked"
            },
            "description": "Filter labels not linked to a product (product_id is null)"
          },
          {
            "name": "order_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Field to sort by",
              "default": "created_at",
              "title": "Order By"
            },
            "description": "Field to sort by"
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Sort direction (asc or desc)",
              "default": "desc",
              "title": "Order"
            },
            "description": "Sort direction (asc or desc)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Page size limit",
              "default": 50,
              "title": "Limit"
            },
            "description": "Page size limit"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Page offset",
              "default": 0,
              "title": "Offset"
            },
            "description": "Page offset"
          },
          {
            "name": "product_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Product type",
              "default": "fertilizer",
              "title": "Product Type"
            },
            "description": "Product type"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LimitOffsetPage_LabelListItem_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/labels/{label_id}/images": {
      "post": {
        "tags": ["labels"],
        "summary": "Create Label Image",
        "description": "Create pending LabelImage record.",
        "operationId": "labels_create_label_image",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "label_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Label Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PresignedUrlRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LabelImageDetail" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["labels"],
        "summary": "Read Label Images",
        "description": "Get label images (without presigned URLs).",
        "operationId": "labels_read_label_images",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "label_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Label Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/LabelImageDetail" },
                  "title": "Response Labels Read Label Images"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/labels/{label_id}/images/{image_id}": {
      "get": {
        "tags": ["labels"],
        "summary": "Read Label Image",
        "description": "Get single label image (without presigned URL).",
        "operationId": "labels_read_label_image",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "image_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Image Id"
            }
          },
          {
            "name": "label_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Label Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LabelImageDetail" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["labels"],
        "summary": "Delete Label Image",
        "description": "Delete a label image, its storage file, and renumber remaining images.",
        "operationId": "labels_delete_label_image",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "image_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Image Id"
            }
          },
          {
            "name": "label_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Label Id"
            }
          }
        ],
        "responses": {
          "204": { "description": "Successful Response" },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/labels/{label_id}/images/complete": {
      "post": {
        "tags": ["labels"],
        "summary": "Complete Label Image Upload",
        "description": "Complete upload by updating LabelImage status from pending to completed.",
        "operationId": "labels_complete_label_image_upload",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "label_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Label Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadCompletionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LabelImageDetail" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/labels/{label_id}/images/{image_id}/presigned-upload-url": {
      "get": {
        "tags": ["labels"],
        "summary": "Get Label Image Presigned Upload Url",
        "description": "Get presigned upload URL for a pending label image.",
        "operationId": "labels_get_label_image_presigned_upload_url",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "image_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Image Id"
            }
          },
          {
            "name": "label_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Label Id"
            }
          },
          {
            "name": "content_type",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Content type (image/png, image/jpeg, image/webp)",
              "title": "Content Type"
            },
            "description": "Content type (image/png, image/jpeg, image/webp)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresignedUploadUrlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/labels/{label_id}/images/{image_id}/presigned-download-url": {
      "get": {
        "tags": ["labels"],
        "summary": "Get Label Image Presigned Download Url",
        "description": "Get presigned download URL for a completed label image.",
        "operationId": "labels_get_label_image_presigned_download_url",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "image_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Image Id"
            }
          },
          {
            "name": "label_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Label Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresignedDownloadUrlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LabelCreate": {
        "properties": {
          "product_type": { "type": "string", "title": "Product Type" },
          "product_id": {
            "anyOf": [
              { "type": "string", "format": "uuid" },
              { "type": "null" }
            ],
            "title": "Product Id"
          }
        },
        "type": "object",
        "required": ["product_type"],
        "title": "LabelCreate"
      },
      "LabelCreated": {
        "properties": {
          "id": { "type": "string", "format": "uuid", "title": "Id" }
        },
        "type": "object",
        "required": ["id"],
        "title": "LabelCreated",
        "description": "Response schema for label creation."
      },
...
      "LabelDataLite": {
        "properties": {
          "brand_name_en": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Brand Name En"
          },
          "brand_name_fr": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Brand Name Fr"
          },
          "product_name_en": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Product Name En"
          },
          "product_name_fr": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Product Name Fr"
          }
        },
        "type": "object",
        "title": "LabelDataLite"
      },
...
      "LabelImageDetail": {
        "properties": {
          "id": { "type": "string", "format": "uuid", "title": "Id" },
          "label_id": {
            "type": "string",
            "format": "uuid",
            "title": "Label Id"
          },
          "display_filename": { "type": "string", "title": "Display Filename" },
          "file_path": { "type": "string", "title": "File Path" },
          "sequence_order": { "type": "integer", "title": "Sequence Order" },
          "status": { "$ref": "#/components/schemas/UploadStatus" },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "current_image_count": {
            "anyOf": [{ "type": "integer" }, { "type": "null" }],
            "title": "Current Image Count"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label_id",
          "display_filename",
          "file_path",
          "sequence_order",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "LabelImageDetail"
      },
      "LabelListItem": {
        "properties": {
          "id": { "type": "string", "format": "uuid", "title": "Id" },
          "review_status": { "$ref": "#/components/schemas/ReviewStatus" },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "label_data": {
            "anyOf": [
              { "$ref": "#/components/schemas/LabelDataLite" },
              { "type": "null" }
            ]
          }
        },
        "type": "object",
        "required": ["id", "review_status", "created_at", "updated_at"],
        "title": "LabelListItem",
        "description": "Schema for label list items."
      },
...
      "LimitOffsetPage_LabelListItem_": {
        "properties": {
          "items": {
            "items": { "$ref": "#/components/schemas/LabelListItem" },
            "type": "array",
            "title": "Items"
          },
          "total": { "type": "integer", "minimum": 0.0, "title": "Total" },
          "limit": { "type": "integer", "minimum": 1.0, "title": "Limit" },
          "offset": { "type": "integer", "minimum": 0.0, "title": "Offset" }
        },
        "type": "object",
        "required": ["items", "total", "limit", "offset"],
        "title": "LimitOffsetPage[LabelListItem]"
      },
...
      "PresignedDownloadUrlResponse": {
        "properties": {
          "presigned_url": { "type": "string", "title": "Presigned Url" }
        },
        "type": "object",
        "required": ["presigned_url"],
        "title": "PresignedDownloadUrlResponse",
        "description": "Response schema for presigned download URL request."
      },
      "PresignedUploadUrlResponse": {
        "properties": {
          "url": { "type": "string", "title": "Url" },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": ["url", "expires_at"],
        "title": "PresignedUploadUrlResponse",
        "description": "Response schema for presigned upload URL request."
      },
      "PresignedUrlRequest": {
        "properties": {
          "display_filename": { "type": "string", "title": "Display Filename" },
          "content_type": {
            "type": "string",
            "enum": ["image/png", "image/jpeg", "image/webp"],
            "title": "Content Type"
          },
          "sequence_order": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Sequence Order"
          }
        },
        "type": "object",
        "required": ["display_filename", "content_type", "sequence_order"],
        "title": "PresignedUrlRequest",
        "description": "Request schema for individual presigned URL generation."
      },
...
      "ReviewStatus": {
        "type": "string",
        "enum": ["not_started", "in_progress", "completed"],
        "title": "ReviewStatus"
      },
...
      "UploadCompletionRequest": {
        "properties": {
          "storage_file_path": {
            "type": "string",
            "title": "Storage File Path"
          }
        },
        "type": "object",
        "required": ["storage_file_path"],
        "title": "UploadCompletionRequest",
        "description": "Request schema for upload completion notification."
      },
      "UploadStatus": {
        "type": "string",
        "enum": ["pending", "completed"],
        "title": "UploadStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": { "scopes": {}, "tokenUrl": "/api/v1/login/access-token" }
        }
      }
    }
  }
}